Restructure relational_scalar test
